refactor: split classifier funcs to semantic segments
