+++ src/model/classifier/classifier_execute_model.py
@@ -1,15 +1,16 @@
 import joblib
 from pandas import DataFrame
 
-from infra.env import DATA_FOLDER_PATH
 from infra.redis_queue import get_redis_queue, QueueName, MAX_JOB_TIMEOUT, get_job
+from model.classifier.classifier_version_model import get_model_path
 from scheme.classifier.classifier_scheme import ClassificationResultItem, ClassificationResult
 from scheme.nomenclature.nomenclature_scheme import JobIdRead
 from util.normalize_name import normalize_name
 
 
 def get_groups_by_items(items: list[str], model_id: str) -> list[ClassificationResultItem]:
-    model = joblib.load(f"{DATA_FOLDER_PATH}/model_{model_id}.pkl")
+    model_path = get_model_path(model_id)
+    model = joblib.load(model_path)
     result: list[ClassificationResultItem] = []
 
     normalized_data = DataFrame({
